moved admin upload to new file
