+++ src/router/index.ts
@@ -1,5 +1,5 @@
 import LandingV from '@/components/landing.vue';
-import HomeV from '@/components/home.vue';
+import HomeV from '@/components/home/home.vue';
 import MetadataEditorV from '@/components/metadata/metadata-editor.vue';
 import StoryPreviewV from '@/components/preview.vue';
 import { createRouter, createWebHashHistory, RouteLocation, RouteLocationNormalized } from 'vue-router';
